+++ desktop/cmp/appOption/ThemeAppOption.ts
@@ -5,8 +5,7 @@
  * Copyright © 2026 Extremely Heavy Industries Inc.
  */
 import {XH} from '@xh/hoist/core';
-import {button} from '@xh/hoist/desktop/cmp/button';
-import {buttonGroupInput, ButtonGroupInputProps} from '@xh/hoist/desktop/cmp/input';
+import {segmentedControl, SegmentedControlProps} from '@xh/hoist/desktop/cmp/input';
 import {Icon} from '@xh/hoist/icon/Icon';
 import {FormFieldProps} from '@xh/hoist/desktop/cmp/form';
 import '@xh/hoist/desktop/register';
@@ -14,8 +13,8 @@ import '@xh/hoist/desktop/register';
 interface ThemeAppOptionSpec {
     /** Props for nested FormField */
     formFieldProps?: FormFieldProps;
-    /** Props for nested ButtonGroupInput */
-    inputProps?: ButtonGroupInputProps;
+    /** Props for nested SegmentedControl */
+    inputProps?: SegmentedControlProps;
 }
 
 /**
@@ -26,16 +25,17 @@ export const themeAppOption = ({formFieldProps, inputProps}: ThemeAppOptionSpec 
         name: 'theme',
         formField: {
             label: 'Theme',
-            item: buttonGroupInput({
-                items: [
-                    button({value: 'light', text: 'Light', icon: Icon.sun(), width: '33.33%'}),
-                    button({value: 'dark', text: 'Dark', icon: Icon.moon(), width: '33.33%'}),
-                    button({value: 'system', text: 'System', icon: Icon.sync(), width: '33.33%'})
+            item: segmentedControl({
+                options: [
+                    {value: 'light', label: 'Light', icon: Icon.sun()},
+                    {value: 'dark', label: 'Dark', icon: Icon.moon()},
+                    {value: 'system', label: 'System', icon: Icon.sync()}
                 ],
                 ...inputProps
             }),
             ...formFieldProps
         },
+        refreshRequired: false,
         prefName: 'xhTheme',
         valueSetter: v => XH.setTheme(v)
     };